% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_issue_content.R
\name{get_issue_content}
\alias{get_issue_content}
\title{Extract the issue title, body and comments from a GitLab issue in markdown.}
\usage{
get_issue_content(url = NULL, project_number, issue_number, .token,
  .api_url)
}
\arguments{
\item{url}{GitLab issue URL.}

\item{project_number}{The project Number. If the \code{url} is \code{NULL}.}

\item{issue_number}{The issue Number. If the \code{url} is \code{NULL}.}

\item{.token}{GitLab access token path you save.}

\item{.api_url}{The IP address, if the GitLab is internal.}
}
\description{
Extract the issue title, body and comments from a GitLab issue in markdown.
}
